# sql: avoid unnecessary allocations during insert/update