Primarily affects benchmarks with secondary indexes.

```
name                          old time/op    new time/op    delta
TrackChoices1_Cockroach-8        272µs ± 4%     270µs ± 2%     ~     (p=0.481 n=10+10)
TrackChoices10_Cockroach-8      72.6µs ± 2%    71.0µs ± 2%   -2.23%  (p=0.000 n=10+10)
TrackChoices100_Cockroach-8     44.8µs ± 1%    42.9µs ± 1%   -4.21%  (p=0.000 n=10+10)
TrackChoices1000_Cockroach-8    42.3µs ± 1%    41.6µs ± 3%   -1.55%   (p=0.028 n=9+10)

name                          old allocs/op  new allocs/op  delta
TrackChoices1_Cockroach-8          388 ± 0%       378 ± 0%   -2.58%  (p=0.000 n=10+10)
TrackChoices10_Cockroach-8         134 ± 0%       123 ± 0%   -8.21%  (p=0.000 n=10+10)
TrackChoices100_Cockroach-8        105 ± 0%        94 ± 0%  -10.48%  (p=0.000 n=10+10)
TrackChoices1000_Cockroach-8       102 ± 0%        91 ± 0%  -10.78%  (p=0.000 n=10+10)
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="35" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/7011)
<!-- Reviewable:end -->