@paperstreet 2016-06-02
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

<details><summary><i>Previously, petermattis (Peter Mattis) wrote…</i></summary>
> #### sql: avoid unnecessary allocations during insert/update
> 
> Primarily affects benchmarks with secondary indexes.
> 
> ```
> name                          old time/op    new time/op    delta
> TrackChoices1_Cockroach-8        272µs ± 4%     270µs ± 2%     ~     (p=0.481 n=10+10)
> TrackChoices10_Cockroach-8      72.6µs ± 2%    71.0µs ± 2%   -2.23%  (p=0.000 n=10+10)
> TrackChoices100_Cockroach-8     44.8µs ± 1%    42.9µs ± 1%   -4.21%  (p=0.000 n=10+10)
> TrackChoices1000_Cockroach-8    42.3µs ± 1%    41.6µs ± 3%   -1.55%   (p=0.028 n=9+10)
> 
> name                          old allocs/op  new allocs/op  delta
> TrackChoices1_Cockroach-8          388 ± 0%       378 ± 0%   -2.58%  (p=0.000 n=10+10)
> TrackChoices10_Cockroach-8         134 ± 0%       123 ± 0%   -8.21%  (p=0.000 n=10+10)
> TrackChoices100_Cockroach-8        105 ± 0%        94 ± 0%  -10.48%  (p=0.000 n=10+10)
> TrackChoices1000_Cockroach-8       102 ± 0%        91 ± 0%  -10.78%  (p=0.000 n=10+10)
> ```
> 
> 
> 
</details>


---

Review status: 0 of 5 files reviewed at latest revision, 3 unresolved discussions, all commit checks successful.

---

*[sql/sqlbase/table.go, line 926 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7011#-KJHDQajUSSxAwL41cMc:-KJHDQajUSSxAwL41cMd:-1003064510) ([raw file](https://github.com/cockroachdb/cockroach/blob/2bd06ecb291347d810dc45533d3578d974123f73/sql/sqlbase/table.go#L926)):*
> ```Go
> func EncodeSecondaryIndex(
> 	tableID ID,
> 	secondaryIndex *IndexDescriptor,
> ```

curiosity: this is because EncodeIndexKey takes a pointer?

---

*[sql/sqlbase/table.go, line 937 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7011#-KJHC_-tPEL899RfD8_N:-KJHC_-tPEL899RfD8_O:-2068272809) ([raw file](https://github.com/cockroachdb/cockroach/blob/2bd06ecb291347d810dc45533d3578d974123f73/sql/sqlbase/table.go#L937)):*
> ```Go
> 	}
> 
> 	// Add the implicit columns - they are encoded ascendingly.
> ```

"encoded ascendingly" is no longer relevant

---

*[sql/sqlbase/table.go, line 973 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7011#-KJHCuqJUQ_hyQ5-G53h:-KJHCuqJUQ_hyQ5-G53i:-1167976611) ([raw file](https://github.com/cockroachdb/cockroach/blob/2bd06ecb291347d810dc45533d3578d974123f73/sql/sqlbase/table.go#L973)):*
> ```Go
> 
> // EncodeSecondaryIndexes encodes key/values for the secondary indexes. colMap
> // maps ColumnIDs to indices in `values`.
> ```

mention that secondaryIndexEntries is the return value and that it's expected to have length = len(indexes)

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/7011#-:-KJHDk6tu7wu3jTEIN0E:1417088721)*
<!-- Sent from Reviewable.io -->