@petermattis 2016-06-02
Review status: 0 of 5 files reviewed at latest revision, 3 unresolved discussions, all commit checks successful.

---

*[sql/sqlbase/table.go, line 926 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7011#-KJHDQajUSSxAwL41cMc:-KJHHdmEd560b4Iafu2C:677251276) ([raw file](https://github.com/cockroachdb/cockroach/blob/2bd06ecb291347d810dc45533d3578d974123f73/sql/sqlbase/table.go#L926)):*
<details><summary><i>Previously, paperstreet (Daniel Harrison) wrote…</i></summary>
> curiosity: this is because EncodeIndexKey takes a pointer?
</details>
Yes.

---

*[sql/sqlbase/table.go, line 937 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7011#-KJHC_-tPEL899RfD8_N:-KJHHmOqsm3vPC1gatsC:-1305696390) ([raw file](https://github.com/cockroachdb/cockroach/blob/2bd06ecb291347d810dc45533d3578d974123f73/sql/sqlbase/table.go#L937)):*
<details><summary><i>Previously, paperstreet (Daniel Harrison) wrote…</i></summary>
> "encoded ascendingly" is no longer relevant
</details>
Well, the comment is still true as passing `nil` for the directions causes ascending order to be used. I'll clarify the comment.

---

*[sql/sqlbase/table.go, line 973 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7011#-KJHCuqJUQ_hyQ5-G53h:-KJHI8gmRy64Gj6g5vF3:-499146009) ([raw file](https://github.com/cockroachdb/cockroach/blob/2bd06ecb291347d810dc45533d3578d974123f73/sql/sqlbase/table.go#L973)):*
<details><summary><i>Previously, paperstreet (Daniel Harrison) wrote…</i></summary>
> mention that secondaryIndexEntries is the return value and that it's expected to have length = len(indexes)
</details>
Done.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/7011)*
<!-- Sent from Reviewable.io -->